moved test_terraform_request_sequence
